removed reply and show thread buttons for comments
Honestly... don't want to have to code that. Getting the id for the parent comment is going to be a hassle... but in case we change our minds then here is the code to add the buttons. I haven't touched the stylesheets concerning this code so just copy and paste this back in and everything should be good.

<div id="comment_buttons">
            <button id="reply_comment" type="button">répondre</button>
            <button id="show_thread" type="button" onClick={toggleThread}>{threadBtnText}</button>
        </div>
        {showThread && (
                <div className="thread">
                    {dummyComments.map((comment,index) => (
                        <Comment
                        key={index}
                        author={comment.author}
                        timestamp={comment.timestamp}
                        content={comment.content} />
                    ))}
                </div>
            )}